feat: auth form with button and tabs
