+++ src/features/auth/form/index.module.scss
@@ -1,13 +1,6 @@
 .auth {
   max-width: 680px;
 
-  // Test tubs
-  &__tabs {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-  }
-
   &__form {
     display: flex;
     gap: 20px;
